+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All important changes to this project will be documented in this file.
 #### Improvements:
 
 * Switched to using Mojang API to get info about versions
+* Added handling of SNBT components in `tellraw` introduced in snapshot `25w02a`
+* Updated dependencies
 
 #### Fixed: